% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{build.MSig2R}
\alias{build.MSig2R}
\title{Save gmt 2 R data file}
\usage{
build.MSig2R(gmt.files, from.tax_id = 9606, to.tax_id = 10090)
}
\arguments{
\item{from.tax_id}{Taxonomy ID of specie of current MSig gmt file}

\item{to.tax_id}{Taxonomy ID of desires}

\item{gmt.file}{Path to gmt file}
}
\value{
Returns a list of pathways and corresponding gene IDs
}
\description{
Function used to save gmt file from the Broad Institute
}
\author{
David Chisanga
}
\keyword{gmt,save,file,broad}
