% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{build.orthologDB}
\alias{build.orthologDB}
\title{Build ortholog DB}
\usage{
build.orthologDB()
}
\value{
The database is written and saved to a R data file and can be read back to R for further inspection or analysis
}
\description{
This function is used to collate gene orthologs between human and mouse genes
from several sources.
}
