% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{change_DE_Status_coef}
\alias{change_DE_Status_coef}
\title{Change DE status}
\usage{
change_DE_Status_coef(
  object,
  contrast.matrix,
  coef = colnames(contrast.matrix)[1],
  filter,
  status
)
}
\arguments{
\item{object}{an RGList, MAList, EList, ExpressionSet or MArrayLM object. Can also be a matrix}

\item{contrast.matrix}{contrast matrix}

\item{coef}{alternative to column for fitted model objects. If specified, then column is ignored.}

\item{filter}{numeric value that is going to be used}

\item{status}{the DE status output from \code{\link[limma:decideTests]{decideTests}}}
}
\value{
A modified DE status object is returned with only the DE status of genes associated with the coef passed
being modified.
}
\description{
This function is used to change the DE status of a feature based on the expression pattern
in the samples used for peforming the comparison
}
