% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{check_gene_history}
\alias{check_gene_history}
\title{A check gene history function}
\usage{
check_gene_history(GeneID, speciesID = 9606, verbose = F)
}
\arguments{
\item{GeneID}{list of gene Ids to check for that are likely to have been discontinued/replaced, has to be EntrezIDs}

\item{speciesID}{Numeric value of any valid species ID such as 9606 for humans and 10090 for mouse. 9606 by default}

\item{verbose}{Logical value. Indicates whether to print to the standard screen output messages}
}
\value{
Returns a vector list of Discontinued and Current GeneIds from the list entered
check.gene.history()
}
\description{
This function checks the history/changes that a given gene ID has undergone
}
\examples{
check.gene.history(c(243187,319599,319785),10090)

}
\author{
David Chisanga
}
\keyword{Genes,}
\keyword{History}
