% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{checkFileAge}
\alias{checkFileAge}
\title{Returns the age of a file in weeks}
\usage{
checkFileAge(file.path)
}
\arguments{
\item{file.path}{String value of the path to the file}
}
\value{
Returns a numeric value of the age of the file in weeks
}
\description{
Returns the age of a file in weeks
}
\examples{
file.create("Example.txt")
checkFileAge("Example.txt")

}
\author{
David Chisanga
}
