% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{convertGeneIds}
\alias{convertGeneIds}
\title{Map gene IDs}
\usage{
convertGeneIds(
  ids.to_map = NULL,
  from = "entrez_id",
  to = "ensembl_gene_id",
  taxo_id = 10090,
  recency = 12
)
}
\arguments{
\item{ids.to_map}{Character vector of IDs to convert}

\item{from}{String indicating the source DB of the IDs, 'entrez_id' by default.
Other accepatable values include: 'ensembl_gene_id','entrez_id','HGNC_id' or 'MGI_id'}

\item{to}{String indicating the desired DB IDs, 'ensembl_gene_id' by default.
Other values include; 'ensembl_gene_id','entrez_id','HGNC_id' or 'MGI_id'}

\item{taxo_id}{Taxonomy ID of species under consideration, 10090 (mouse) by default.
Current supported taxonomies include 10090 (mouse) and 9606 (human)}
}
\value{
A list of the mapping results is returned.
}
\description{
Function to map gene IDs between Ensembl and EntrezIDs
Uses the gene2ensembl mapping from NCBI to map gene IDs between the 2 annotations
}
\examples{
convertGeneIds(ids.to_map = c("16640","16523","14938","13051","94226","17059"))


ids.convert=c("ENSMUSG00000033024","ENSMUSG00000030247",
"ENSMUSG00000023132","ENSMUSG00000052336",
"ENSMUSG00000045087", "ENSMUSG00000030325")

convertGeneIds(ids.to_map=ids.convert,from="ensembl_gene_id",to="entrez_id")
}
\author{
David Chisanga
}
