% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{download.geneInfo}
\alias{download.geneInfo}
\title{Function to download gene info files}
\usage{
download.geneInfo(
  species = "Hs",
  database = "RefSeq",
  version = 94,
  file.age = 3
)
}
\arguments{
\item{species}{Character symbol of species to download, default is Hs. Use Mm for mouse and}

\item{database}{String name of database to download from. RefSeq by default. Use Ensembl}

\item{version}{Numeric of the version  of the database to download from}

\item{file.age}{Number representing the age of the file before it can be updated}
}
\description{
Function to download gene info files
}
