% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{downloadGeneAnnotations}
\alias{downloadGeneAnnotations}
\title{Download gene annotations}
\usage{
downloadGeneAnnotations()
}
\value{
Returns a dataframe
}
\description{
Function is used to download and merge 3 gene annotations for mice and human to create a
database that is then used in converting gene IDs
}
\author{
David Chisanga
}
