% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{generateKableTables}
\alias{generateKableTables}
\title{A function to generate kable tables in a Knitr Rmd.}
\usage{
generateKableTables(
  table.data,
  caption = NULL,
  long.table = F,
  format.args = list(big.mark = ","),
  booktabs = T,
  full.width = F,
  latex.options = switch("basic", basic = c("basic", "HOLD_position"), scale =
    c("scale_down", "HOLD_position"), repeat_header = c("HOLD_position", "repeat_header")),
  row.names = F,
  digits = 2,
  align = NULL,
  fontsize = NULL,
  format = "latex",
  escape = T,
  kable_style.args = NULL,
  ...
)
}
\arguments{
\item{table.data}{a matrix/data frame of rows and columns to be displayed in the table}

\item{caption}{title/caption to shown on table}

\item{long.table}{logic value to show long format table}

\item{format.args}{list, see ?kable}

\item{full.width}{logic value, full width of table}

\item{latex.options}{list of options, see ?kable_styling}

\item{row.names}{logic value, used to display rown names of matrix/data frame}

\item{digits}{numeric value, see ?kable for details}

\item{align}{see ?kable_styling}

\item{kable_style.args}{list of extra parameters to be passed to the kable_styling function. See ?kable_styling}

\item{...}{extra parameters are passed to the kableExtra::kable function, see ?kableExtra::kable}

\item{font_size}{numeric for font size}
}
\value{
Returns a kable object
}
\description{
A function to generate kable tables in a Knitr Rmd.
}
\note{
Only valid for latex style documents.
}
\examples{
data=data.frame(X=1:4,Y=1:4,Z=2*(1:4))
generateKableTables(data)
}
\author{
David Chisanga
}
