% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{get.inbuilt.MSigs}
\alias{get.inbuilt.MSigs}
\title{Function to load Molecular signature databases}
\usage{
get.inbuilt.MSigs(collection = NULL, species = "Mm", show.version = T)
}
\arguments{
\item{collection}{character or list of collections to be loaded, by default is NULL meaning all collections are loaded.
Options include, "c1","c2","c3","c4","c5","c6","c7" or h or a combined list of these}

\item{species}{Character string of the species, currently supports human (hs) and mouse (mm)}
}
\value{
Returns a list of paths to the inbuilt collections
}
\description{
The function is used to load molecular signature databases provided by the Broad Institute.
When run for the first time, the function first
}
\author{
David Chisanga
}
