% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{get.orthologs}
\alias{get.orthologs}
\title{Map gene IDs between species}
\usage{
get.orthologs(
  from.taxid = 9606,
  to.taxid = 10090,
  verbose = F,
  ortho_file.recency = 12,
  Entrez.IDs = NULL
)
}
\arguments{
\item{from.taxid}{Numeric taxonomy ID of the specie for which the GeneIDs are being converted from}

\item{to.taxid}{Numeric taxonomy ID of the specie for which the GeneIDs are being converted to}

\item{verbose}{Logic, display message if True and vice-versa}
}
\value{
Returns a list with the mapping between Gene IDs from the 2 species
}
\description{
Function to convert Entrez Gene IDs between species
}
