% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{get_geneIDs}
\alias{get_geneIDs}
\title{Map symbols to gene IDs}
\usage{
get_geneIDs(
  database = "RefSeq",
  gene.symbols = NULL,
  specie = "Mm",
  ignore.case = F
)
}
\arguments{
\item{database}{a string value of the database to search against, accepted values include; 'RefSeq', 'Ensembl' and 'GENCODE','RefSeq' by default}

\item{gene.symbols}{list of gene symbols to look up, NULL by default}

\item{specie}{a string value of the species to look up, accepted values are 'Mm' for Mouse and 'Hs' for Homo sapiens, 'Mm' by default}

\item{ignore.case}{logic value indicates whether or not to ignore the case of the symbols during the search, FALSE by default}
}
\value{
Returns a list with the input gene symbols as keys and their associated details including, GeneID, Symbol and Synonyms.
}
\description{
The function maps gene symbols to gene IDs
}
\examples{
rst=get.geneIDs(gene.symbols = c("mcad","Kras"),specie = "Hs",ignore.case = T)
}
\author{
David Chisanga
}
\keyword{EnsemblID}
\keyword{EntrezID,}
\keyword{Gene}
\keyword{Genes,}
\keyword{Symbol,}
