% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{getGeneInfoEnsembl}
\alias{getGeneInfoEnsembl}
\title{Get Ensembl gene information}
\usage{
getGeneInfoEnsembl(
  species = "Mm",
  info.details = NULL,
  verbose = F,
  version = 94
)
}
\arguments{
\item{species}{String indicating for which gene info is required, values can 'Hs' or 'Mm'. 'Mm' by default}

\item{info.details}{Character list of columns/details to return. All columns are returned by default}

\item{verbose}{Logical value, enables the printout of messages during downloads}

\item{version}{Numeric value for the version of Ensembl, 94 by default}
}
\value{
Returns a dataframe with rows corresponding to genes
}
\description{
Function retrieves the Gene info details for either Humans or Mouse from Ensembl
}
\examples{

#Without specifying the species
getGeneInfoEnsembl()

#Specifying the species
getGeneInfoEnsembl(species='Hs')

#Specifying the info details
getGeneInfoEnsembl(species='Hs',info.details=c("gene_id","gene_name","gene_biotype"))
}
\author{
David Chisanga
}
