% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{getGeneInfoNCBI}
\alias{getGeneInfoNCBI}
\title{Get NCBI RefSeq gene information}
\usage{
getGeneInfoNCBI(species = "Mm", info.details = NULL, verbose = F)
}
\arguments{
\item{species}{String indicating for which gene info is required, values can 'Hs' or 'Mm'. 'Mm' by default}

\item{info.details}{Character list of columns/details to return. All columns are returned by default}

\item{verbose}{Logical value, enables the printout of messages during downloads}
}
\value{
Returns a dataframe with rows corresponding to genes
}
\description{
The function is used to retrieve gene information for Humans or Mouse genes
}
\examples{

#Without specifying the species
getGeneInfoNCBI()

#Specifying the species
getGeneInfoNCBI(species='Hs')

#Specifying the info details
getGeneInfoNCBI(species='Hs',info.details=c("GeneID","Symbol","description"))
}
\author{
David Chisanga
}
