% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{getLabelCoordinates}
\alias{getLabelCoordinates}
\title{Generate label coordinates}
\usage{
getLabelCoordinates(
  axis = "y-axis",
  max.distance = 6,
  arrow.coordinates,
  label.coordinates,
  minDist
)
}
\arguments{
\item{axis}{String of axis on which labels/arrows are shown, values include x-axis or y-axis. y-axis by default}

\item{max.distance}{numeric, the maximum distance from the 0 coordinate to place labels, 11 by default}

\item{arrow.coordinates}{list of numeric coordinates on the axis on which labels will be placed}

\item{label.coordinates}{list of numeric coordinates on the other axis for labels}

\item{minDist}{The minimum expected distance between near labels}
}
\value{
Returns a list of spaced coordinates
}
\description{
A function to space labels on a plot with Arrows, usually MDS plots and Volcano plot
}
\author{
David Chisanga
}
