% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{labelMDplot}
\alias{labelMDplot}
\title{Wrapper function to include arrows to a MD plot in limma}
\usage{
labelMDplot(
  x,
  y,
  min.dist = 0.4,
  y.inset = 0.8,
  x.inset = 0.8,
  symbols,
  cex.lbl = 0.7,
  arrow.angle = 30,
  text.srt = 90,
  col.lbl = "black",
  axis = "y-axis",
  col.arrow = "black",
  max.distance = 7
)
}
\arguments{
\item{x}{numeric coordinate of arrow on the x-axis}

\item{y}{numeric coordiate of arrow on the y-axis}

\item{min.dist}{The minimum expected distance between near labels}

\item{y.inset}{value to extend the start of the y-coordinate of label}

\item{x.inset}{value to extend the start of the x-coordinate of label}

\item{symbols}{character of gene symbols to be used in labelling}

\item{arrow.angle}{angle of arrows}

\item{text.srt}{angle rotation for labels}

\item{col.lbl}{character vector of colors for labels}

\item{cex}{size of labels}
}
\description{
Wrapper function to include arrows to a MD plot in limma
}
\author{
David Chisanga
}
