% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{outersect}
\alias{outersect}
\title{Function to get the elements in set a that are not in set b}
\usage{
outersect(set.a, set.b)
}
\arguments{
\item{set.a}{vector of elements to compare}

\item{set.b}{vector of elements to compare}
}
\value{
Returns a vector of values not found in set b
}
\description{
Function to get the elements in set a that are not in set b
}
\author{
David Chisanga
}
