% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{plot_DEVenn}
\alias{plot_DEVenn}
\title{Plot venn diagram of DE genes}
\usage{
plot_DEVenn(
  status,
  inset = c(0.19, 0.1),
  de.cols,
  include = c("up", "down"),
  leg.bg = "white",
  leg.box_col = "white",
  leg.cex = 0.6,
  ...
)
}
\arguments{
\item{status}{DE status output from \code{\link[limma:decideTests]{decideTests}}}

\item{inset}{numeric vector of values used to place the legend on the venn diagram}

\item{de.cols}{vector of colors used to represent the DE status}

\item{include}{vector used to represent the DE status}

\item{leg.bg}{background color of the legend}

\item{leg.box_col}{box color for the legend}

\item{leg.cex}{size for legend text}

\item{...}{all other parameters are passed to \code{\link[limma:vennDiagram]{vennDiagram}}}
}
\value{
A Venn diagram is plotted to the graphics object
}
\description{
Wrapper function to plot the Venn diagram generated by the \code{\link[limma:vennDiagram]{vennDiagram}}
}
