% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{plotbarcode}
\alias{plotbarcode}
\title{Modified barcodeplot function}
\usage{
plotbarcode(
  statistics,
  index = NULL,
  index2 = NULL,
  gene.weights = NULL,
  weights.label = "Weight",
  labels = c("Down", "Up"),
  quantiles = c(-1, 1) * sqrt(2),
  col.bars = NULL,
  alpha = 0.4,
  worm = TRUE,
  span.worm = 0.45,
  xlab = "Statistic",
  split.worm = T,
  col.uenrich = "pink",
  col.denrich = "lightblue",
  enrichment.height = NULL,
  ...
)
}
\description{
Modified barcodeplot function
}
