% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{plotDensities}
\alias{plotDensities}
\title{Density plot}
\usage{
plotDensities(
  raw.counts,
  log.base = 2,
  main = "Density plot of raw counts",
  cex.legend = 0.5,
  cex.lab = 0.9,
  cex.axis = 0.8,
  line.cols = NULL,
  legend = T,
  y.lim = NULL,
  ...
)
}
\arguments{
\item{raw.counts}{Matrix/Dataframe of raw counts}

\item{log.base}{Numeric value of the base of the log to be used. 2 by default}

\item{main}{Title of plot}

\item{cex.legend}{Numeric value of the text size on the legend}

\item{cex.lab}{Numeric value of the size of the axis labels}

\item{cex.axis}{Numeric value of the size of the tick-labels}

\item{legend}{Logic value to display the legend}

\item{cex.main}{Numeric value of the size of the title}
}
\value{
Generates plot of density values for each column (sample) and returns a list of densities
}
\description{
Function to generate a multi-density plot
}
\examples{
raw.counts<-matrix(rnbinom(100,size=1,mu=10),25,4)
plotDensities(raw.counts)
plotDensities(raw.counts,10)
}
\author{
David Chisanga
}
