% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{plotMD.dc}
\alias{plotMD.dc}
\title{Function that emulates the plotMD function in limma}
\usage{
plotMD.dc(
  object,
  column = ncol(object),
  coef = NULL,
  cex.main = 1,
  xlab = "Average log-expression",
  include.abline = T,
  ylab = "logFC",
  cex.lbl = 0.6,
  angle = 30,
  xlim = NULL,
  ylim = NULL,
  cex = 0.3,
  de.cols = c("red", "blue"),
  include.labels = T,
  label.only = NULL,
  main = column,
  status,
  label.ids = NULL,
  max.distance = NULL,
  minDist = 0.4,
  arrow.col = "black",
  label.offset = 0.2,
  srt.lbl = 20,
  ...
)
}
\arguments{
\item{object}{an RGList, MAList, EList, ExpressionSet or MArrayLM object.}

\item{column}{integer, column of object to be plotted.}

\item{coef}{alternative to column for fitted model objects. If specified, then column is ignored.}

\item{xlab}{String, x-axis label}

\item{include.abline}{logical, include horizontal line}

\item{ylab}{String, y-axis label}

\item{xlim}{numeric vector, limits on the x-axis}

\item{ylim}{numeric vector, limits on the y-axis}

\item{de.cols}{vector list of colours used for highlght up-regulated and down-regulated genes}

\item{...}{extra parameters are passed to the}
}
\description{
This function is a wrapper function to the plotMD \code{\link[limma:plotMD]{plotMD}} function in \code{\link[limma]{limma}} but adds
the ability to highlight selected genes or other features of interest
}
