% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{plotVolcalno}
\alias{plotVolcalno}
\title{plot volcano}
\usage{
plotVolcalno(
  top.matrix,
  status,
  title = "Volcano plot",
  ylim = c(0, 11),
  xlim = c(-10, 10),
  label.Ids = NA,
  cex.lbl = 0.6,
  minDist = 0.25,
  use.raw = F,
  cex.points = 0.6,
  angle = 30,
  arrow.col = "black",
  key.size = 1.5,
  de.cols = c("red", "blue")
)
}
\arguments{
\item{top.matrix}{A matrix retained by either \code{\link[limma:topTable]{topTable}} or \code{\link[limma:topTreat]{topTreat}}}

\item{status}{results from decideTest with the DE status}

\item{title}{Text to be used for the title of the plot}

\item{ylim}{limits on the y-axis}

\item{xlim}{limits on the x-axis}

\item{label.Ids}{gene IDs for the genes to be used for highlighting genes on the plot}

\item{cex.lbl}{plot size for the highlighted text}

\item{minDist}{minimum distance between text used to determine the proximity of highlighted text}

\item{cex.points}{minimum plot size for the points}

\item{angle}{angle used to rotate the arrows}

\item{arrow.col}{colour used for the arrows.}

\item{key.size}{factor to multiply the minimum plot size for DE points}

\item{de.cols}{vector list of colours used for highlght up-regulated and down-regulated genes}
}
\description{
Function to generate volcano plot
}
\author{
David Chisanga
}
