% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{read.gmt}
\alias{read.gmt}
\title{Read gmt file}
\usage{
read.gmt(gmt.file)
}
\arguments{
\item{gmt.file}{Path to gmt file}
}
\value{
Returns a list of pathways and corresponding gene IDs
}
\description{
Function used to read in gmt file from the Broad Institute
}
\author{
David Chisanga
}
\keyword{gmt,read,file,broad}
