% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{spaceLabels}
\alias{spaceLabels}
\title{Space text labels}
\usage{
spaceLabels(
  labelCoords,
  minDist = 0.4,
  boundaries = c(0, max(labelCoords) + 2)
)
}
\arguments{
\item{labelCoords}{a numeric vector list of coordinates}

\item{minDist}{The minimum expected distance between near labels}
}
\value{
Returns a vector of spaced coordinates for labels
}
\description{
This function is used to space multiple text labels
}
\examples{
labels=list(seq(1,3,0.3))
names(labels)=paste0("Label0000",1:length(labels))
spaceLabels(labels,0.5)

}
\author{
David Chisanga
}
