% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotFunctions.R
\name{spaceLabels_old}
\alias{spaceLabels_old}
\title{Include space between labels}
\usage{
spaceLabels_old(labelCoords, minDist = 0.4)
}
\arguments{
\item{labelCoords}{a numeric vector list of coordinates}

\item{minDist}{The minimum expected distance between near labels}
}
\value{
Returns a vector of spaced coordinates for labels
}
\description{
A function to space labels on a plot with Arrows, usually MDS plots and Volcano plot
}
\examples{
labels=list(seq(1,3,0.3))
names(labels)=paste0("Label0000",1:length(labels))
spaceLabels(labels,0.5)
}
\author{
David Chisanga
}
